rename ExceptionHandler namespace to "Diagnostics"
