+++ lib/Diagnostics/Views/ExceptionStyle.phtml
