+++ lib/Diagnostics/Views/ExceptionView.phtml
