+++ lib/Hosting/WebHost.php
@@ -13,7 +13,7 @@ use DevNet\System\Configuration\IConfiguration;
 use DevNet\System\Dependency\IServiceProvider;
 use DevNet\System\Dependency\ServiceCollection;
 use DevNet\System\Runtime\LauncherProperties;
-use DevNet\Web\Exception\ExceptionHandlerMiddleware;
+use DevNet\Web\Diagnostics\ExceptionHandlerMiddleware;
 use DevNet\Web\Http\HttpContext;
 use DevNet\Web\Http\HttpContextFactory;
 use DevNet\Web\Middleware\IApplicationBuilder;
